Improve how deferrals are tracked
Now, we have three types of data:

1. Items: remotely retrieved information about each URL.
2. Metadata: locally generated information about each URL.
3. Impact scores: regenerated at will, and not persisted.

When `mq up` is executed, it changes (1) but not (2).
When a deferral occurs, it changes (2) but not (1).

(1) and (2) are saved into two separate JSON files in the monoqueue data
folder, so that the entire collection of action item data does not need
to be written out to disk every time a deferral operation happens.