Display GitHub issue body below the impact report
For now, we hardcode item['issue']['body'], because the firefox and
discourse handlers do not currently deliver any body-like text.
If that changes in the future, we can generalize the body field to
be handler-agnostic, and unhardcode from the curses UI.

The body is lightly processed to wrap intelligently at word boundaries.
The string writing routine in general is enhanced to put ellipses when
asked to write strings that are too long for the screen width.

New key handlers allow scrolling the body, when there are too many lines
to fit on the screen at once: body scan be scrolled one line at a time
with J/K or the up/down arrows, or scrolled half a screen at a time with
the PgUp/PgDn keys. There is currently a small bug in that PgUp/PgDn
always scroll 1/2 the entire screen height, rather than 1/2 the body
viewport height, but I'm too lazy to address it for the moment.

Finally, the curses UI is more resilient to window resizing.
Previously, it would occasionally crash when resized to smaller widths.
Unfortunately, the curses.COLS and curses.LINES constants do not change
when the window is resized; there is probably a way to detect and
address resizes, but I did not yet investigate how to do it.